App now uses ExoPlayer instead of Media player
- App now correctly runs in background
- Notification is now separated into 2:
  - 1: Android's default media player notification for controlling playback
  - 2: Small notification that shows lyrics if available
 - Ui improvements

+++ app/src/main/java/mb28/crysongs/ui/fullscreen_player/LyricsTab.kt
@@ -117,7 +117,7 @@ private fun LyricText(i: Int, state: LazyListState, scope: CoroutineScope) {
                             lerp(
                                 0, duration,
                                 inverseLerp(0f, duration / 1000f, line.TimeStomp),
-                            )
+                            ).toLong()
                         )
                         state.scrollToItem(i, -500)
                     }
